+++ layouts/text/plugin.php
@@ -1,41 +1,21 @@
 <?php
 
 namespace DustPress\Components;
+use Geniem\ACF\Field;
 
 class Text extends Component {
 	var $label = 'Text';
 	var $name = 'text';
 
 	public function fields() {
-		return array (
-			'key' => 'dpc_text',
-			'name' => $this->name,
-			'label' => $this->label,
-			'display' => 'block',
-			'sub_fields' => array (
-				array (
-					'key' => 'dpc_text_text',
-					'label' => 'Text',
-					'name' => 't',
-					'type' => 'textarea',
-					'instructions' => '',
-					'required' => 0,
-					'conditional_logic' => 0,
-					'wrapper' => array (
-						'width' => '',
-						'class' => '',
-						'id' => '',
-					),
-					'default_value' => '',
-					'placeholder' => '',
-					'prepend' => '',
-					'append' => '',
-					'maxlength' => '',
-				),
-			),
-			'min' => '',
-			'max' => '',
-		);
+		return (new Field\Flexible\Layout( $this->name ))
+			->set_key( 'dpc_text' )
+			->set_name( $this->name )
+			->set_label( $this->label )
+			->add_field( (new Field\Textarea( 'Text' ))
+				->set_key( 'dpc_text_text' )
+				->set_name( 't' )
+			);
 	}
 
 	public function data( $data ) {