+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.6.2] - 2017-10-30
+
 ### Added
 - Added ACF Codifier as a dependency for the plugin
 - Made component's data function to be overridable